Move gossip error to a separate file and fix linting
