+++ packages/lodestar/src/network/gossip/errors.ts
@@ -1,0 +1,7 @@
+export class GossipValidationError extends Error {
+  public code: string;
+  constructor(code: string, message?: string) {
+    super(message);
+    this.code = code;
+  }
+}
